task3/exporter/requirements.txt
grpcio==1.47.0
grpcio-tools==1.47.0
protobuf==3.20.1
pyyaml==6.0